change passphrase support; add cops scripts

+++ egs/babel/s5c/mysteps/make_mfcc.sh
@@ -1,6 +1,9 @@
-#!/bin/bash 
+#!/bin/bash
+{
+set -e
+set -o pipefail
 
-# Copyright 2012  Johns Hopkins University (Author: Daniel Povey)
+# Copyright 2012-2016  Johns Hopkins University (Author: Daniel Povey)
 # Apache 2.0
 # To be run from .. (one directory up from here)
 # see ../run.sh for example
@@ -10,6 +13,8 @@ nj=4
 cmd=run.pl
 mfcc_config=conf/mfcc.conf
 compress=true
+write_utt2num_frames=false  # if true writes utt2num_frames
+passphrase=
 # End configuration section.
 
 echo "$0 $@"  # Print the command line for logging
@@ -17,19 +22,29 @@ echo "$0 $@"  # Print the command line for logging
 if [ -f path.sh ]; then . ./path.sh; fi
 . parse_options.sh || exit 1;
 
-if [ $# != 3 ]; then
-   echo "usage: make_mfcc.sh [options] <data-dir> <log-dir> <path-to-mfccdir>";
-   echo "options: "
+if [ $# -lt 1 ] || [ $# -gt 3 ]; then
+   echo "Usage: $0 [options] <data-dir> [<log-dir> [<mfcc-dir>] ]";
+   echo "e.g.: $0 data/train exp/make_mfcc/train mfcc"
+   echo "Note: <log-dir> defaults to <data-dir>/log, and <mfccdir> defaults to <data-dir>/data"
+   echo "Options: "
    echo "  --mfcc-config <config-file>                      # config passed to compute-mfcc-feats "
    echo "  --nj <nj>                                        # number of parallel jobs"
    echo "  --cmd (utils/run.pl|utils/queue.pl <queue opts>) # how to run jobs."
+   echo "  --write-utt2num-frames <true|false>     # If true, write utt2num_frames file."
    exit 1;
 fi
 
 data=$1
-logdir=$2
-mfccdir=$3
-
+if [ $# -ge 2 ]; then
+  logdir=$2
+else
+  logdir=$data/log
+fi
+if [ $# -ge 3 ]; then
+  mfccdir=$3
+else
+  mfccdir=$data/data
+fi
 
 # make $mfccdir an absolute pathname.
 mfccdir=`perl -e '($dir,$pwd)= @ARGV; if($dir!~m:^/:) { $dir = "$pwd/$dir"; } print $dir; ' $mfccdir ${PWD}`
@@ -56,39 +71,51 @@ for f in $required; do
     exit 1;
   fi
 done
-myutils/validate_data_dir.sh --no-text --no-feats $data || exit 1;
-
-# note: in general, the double-parenthesis construct in bash "((" is "C-style
-# syntax" where we can get rid of the $ for variable names, and omit spaces.
-# The "for" loop in this style is a special construct.
+utils/validate_data_dir.sh --no-text --no-feats $data || exit 1;
 
 if [ -f $data/spk2warp ]; then
   echo "$0 [info]: using VTLN warp factors from $data/spk2warp"
   vtln_opts="--vtln-map=ark:$data/spk2warp --utt2spk=ark:$data/utt2spk"
+elif [ -f $data/utt2warp ]; then
+  echo "$0 [info]: using VTLN warp factors from $data/utt2warp"
+  vtln_opts="--vtln-map=ark:$data/utt2warp"
 fi
+
+for n in $(seq $nj); do
+  # the next command does nothing unless $mfccdir/storage/ exists, see
+  # utils/create_data_link.pl for more info.
+  utils/create_data_link.pl $mfccdir/raw_mfcc_$name.$n.ark
+done
+
+
+if $write_utt2num_frames; then
+  write_num_frames_opt="--write-num-frames=ark,t:$logdir/utt2num_frames.JOB"
+else
+  write_num_frames_opt=
+fi
+
 
 if [ -f $data/segments ]; then
   echo "$0 [info]: segments file exists: using that."
 
   split_segments=""
-  for ((n=1; n<=nj; n++)); do
+  for n in $(seq $nj); do
     split_segments="$split_segments $logdir/segments.$n"
   done
- 
-  utils/split_scp.pl $data/segments $split_segments || exit 1;
-  rm $logdir/.error 2>/dev/null
+
+  utils/split_scp.pl $data/segments $split_segments
+  [ -f $logdir/.error ] && rm $logdir/.error 2>/dev/null
 
   $cmd JOB=1:$nj $logdir/make_mfcc_${name}.JOB.log \
-    extract-segments scp,p:$scp $logdir/segments.JOB ark:- \| \
+    extract-segments "scp,p:sed 's#PASSPHRASE#$passphrase#' $scp |" $logdir/segments.JOB ark:- \| \
     compute-mfcc-feats $vtln_opts --verbose=2 --config=$mfcc_config ark:- ark:- \| \
-    copy-feats --compress=$compress ark:- \
-      ark,scp:$mfccdir/raw_mfcc_$name.JOB.ark,$mfccdir/raw_mfcc_$name.JOB.scp \
-     || exit 1;
+    copy-feats --compress=$compress $write_num_frames_opt ark:- \
+      ark,scp:$mfccdir/raw_mfcc_$name.JOB.ark,$mfccdir/raw_mfcc_$name.JOB.scp
 
 else
   echo "$0: [info]: no segments file exists: assuming wav.scp indexed by utterance."
   split_scps=""
-  for ((n=1; n<=nj; n++)); do
+  for n in $(seq $nj); do
     split_scps="$split_scps $logdir/wav_${name}.$n.scp"
   done
 
@@ -101,7 +128,7 @@ else
   $cmd JOB=1:$nj $logdir/make_mfcc_${name}.JOB.log \
     compute-mfcc-feats  $vtln_opts --verbose=2 --config=$mfcc_config \
      scp,p:$logdir/wav_${name}.JOB.scp ark:- \| \
-      copy-feats --compress=$compress ark:- \
+      copy-feats $write_num_frames_opt --compress=$compress ark:- \
       ark,scp:$mfccdir/raw_mfcc_$name.JOB.ark,$mfccdir/raw_mfcc_$name.JOB.scp \
       || exit 1;
 fi
@@ -114,14 +141,21 @@ if [ -f $logdir/.error.$name ]; then
 fi
 
 # concatenate the .scp files together.
-for ((n=1; n<=nj; n++)); do
+for n in $(seq $nj); do
   cat $mfccdir/raw_mfcc_$name.$n.scp || exit 1;
-done > $data/feats.scp
+done > $data/feats.scp || exit 1
+
+if $write_utt2num_frames; then
+  for n in $(seq $nj); do
+    cat $logdir/utt2num_frames.$n || exit 1;
+  done > $data/utt2num_frames || exit 1
+  rm $logdir/uttnum_frames.*
+fi
 
 rm $logdir/wav_${name}.*.scp  $logdir/segments.* 2>/dev/null
 
-nf=`cat $data/feats.scp | wc -l` 
-nu=`cat $data/utt2spk | wc -l` 
+nf=`cat $data/feats.scp | wc -l`
+nu=`cat $data/utt2spk | wc -l`
 if [ $nf -ne $nu ]; then
   echo "It seems not all of the feature files were successfully processed ($nf != $nu);"
   echo "consider using utils/fix_data_dir.sh $data"
@@ -133,3 +167,4 @@ if [ $nf -lt $[$nu - ($nu/20)] ]; then
 fi
 
 echo "Succeeded creating MFCC features for $name"
+}
